+++ hotel/settings.py
@@ -38,7 +38,9 @@ INSTALLED_APPS = [
     'django.contrib.sessions',
     'django.contrib.messages',
     'django.contrib.staticfiles',
-    'hotelapp',  
+    'principal', 
+    'funcionarios',
+    'quartos'
 ]
 
 MIDDLEWARE = [
@@ -123,4 +125,4 @@ STATIC_URL = 'static/'
 # https://docs.djangoproject.com/en/5.1/ref/settings/#default-auto-field
 
 DEFAULT_AUTO_FIELD = 'django.db.models.BigAutoField'
-AUTH_USER_MODEL = 'hotelapp.Funcionario'
+AUTH_USER_MODEL = 'funcionarios.Funcionario'
