+++ quartos/templates/tipo/tipos.html
